python-worker/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery[redis]==5.3.6
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
kafka-python==2.0.2
aiokafka==0.8.1
grpcio==1.60.0
grpcio-tools==1.60.0
milvus-client==1.0.1
openai==1.12.0
anthropic==0.18.0
tenacity==8.2.3
redis==5.0.1
python-json-logger==2.0.7
structlog==24.1.0
duckduckgo-search==5.0
sentence-transformers==2.5.0
asyncpg==0.29.0
